Fix support rails 4.2.1
