+++ lib/ridgepole.rb
@@ -13,7 +13,6 @@ require 'ridgepole/diff'
 require 'ridgepole/dsl_parser'
 require 'ridgepole/dumper'
 require 'ridgepole/execute_expander'
-require 'ridgepole/ext/foreign_key'
 require 'ridgepole/logger'
 require 'ridgepole/migration_ext'
 require 'ridgepole/schema_dumper_ext'
