+++ ridgepole.gemspec
@@ -18,15 +18,11 @@ Gem::Specification.new do |spec|
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ['lib']
 
-  spec.add_dependency 'activerecord', '< 4.2.0'
+  spec.add_dependency 'activerecord', '~> 4.2.1'
   spec.add_development_dependency 'bundler'
   spec.add_development_dependency 'rake'
   spec.add_development_dependency 'rspec', '>= 3.0.0'
   spec.add_development_dependency 'mysql2'
   spec.add_development_dependency 'coveralls'
-  spec.add_development_dependency 'activerecord-mysql-unsigned', '~> 0.3.1'
-  spec.add_development_dependency 'foreigner', '<= 1.7.1'
-  spec.add_development_dependency 'activerecord-mysql-pkdump', '>= 0.1.0'
-  spec.add_development_dependency 'migration_comments'
-  spec.add_development_dependency 'activerecord-mysql-awesome', '>= 0.0.3'
+  spec.add_development_dependency 'activerecord-mysql-awesome', '>= 0.0.7'
 end
